* commit 'a7ccebeafb7d3c09b11308f79a13a056a0ff2c2a':
  smoke comes out of base when it loses half of its health

+++ src/text.cpp
@@ -83,7 +83,7 @@ void Text::loadGlyphs()
       texture,
       glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
       glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-      face->glyph->advance.x
+      (unsigned int)face->glyph->advance.x
     };
     Characters.insert(std::pair<GLchar, Character>(c, character));
   }
